+++ package.json
@@ -50,7 +50,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
-    "date-fns": "^4.1.0",
+    "date-fns": "^2.30.0",
     "embla-carousel-react": "^8.6.0",
     "framer-motion": "^12.12.1",
     "inngest": "^3.37.0",
@@ -66,7 +66,7 @@
     "qrcode": "^1.5.4",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.3.0",
-    "react-day-picker": "^9.7.0",
+    "react-day-picker": "^8.10.1",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.3.8",
     "react-hook-form": "^7.56.4",
