+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let i,s={};const e=(e,n)=>(e=new URL(e+".js",n).href,s[e]||new Promise((s=>{if("document"in self){const i=document.createElement("script");i.src=e,i.onload=s,document.head.appendChild(i)}else i=e,importScripts(e),s()})).then((()=>{let i=s[e];if(!i)throw new Error(`Module ${e} didn’t register its module`);return i})));self.define=(n,a)=>{const t=i||("document"in self?document.currentScript.src:"")||location.href;if(s[t])return;let c={};const r=i=>e(i,t),d={module:{uri:t},exports:c,require:r};s[t]=Promise.all(n.map((i=>d[i]||r(i)))).then((i=>(a(...i),c)))}}define(["./workbox-4754cb34"],(function(i){"use strict";importScripts(),self.skipWaiting(),i.clientsClaim(),i.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"a050c28bfac2853b751788b6cf333742"},{url:"/_next/static/chunks/1227-f75e71f3f477e347.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/1532-b70d4b708e219c51.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/1578-8d03de0caab34e24.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/1588-6a2e8de9e4dadbd2.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/164f4fb6-9974aec7cfd4ddc7.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/1684-02e7a30568b3eba4.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/2121.7f3788615384e443.js",revision:"7f3788615384e443"},{url:"/_next/static/chunks/2206-4f77b797139ea24b.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/2290-d2e534377287b673.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/2470-fa8973d56c420687.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/2491-ff27243db1a59847.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/254-6453f02386c64b39.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/3013-953e6cae54d6d37b.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/3063-14eeb2a1e0bd7180.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/3113-578ce651b18f9e60.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/3398-3317ed2107e09593.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/3815-0da0630e5c90c8fc.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/4014-697f66bfdf6244b9.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/405-d93f90d64256b510.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/4543-da08a75a9505d116.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/4610-5ec10466e932ec9a.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/4bd1b696-49d338a46d558a00.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/5181-6cf679e2f1f11d74.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/5273-7cf9ce43ed71779e.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/5452-9752e90b8e594679.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/5493-48ab61f32b1bd8c4.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/5789-1b31edc9f1d9534d.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/5825-21ed8db0b7db527d.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/602dbae6-939090e577254594.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/6072-726b3c6df065048d.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/6213-57497bf22762688b.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/6350-2941f0a8e089e9e3.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/6413-9bc8cf3933107103.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/6656-0d2bdd4bec4e1c8c.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/6671-69e0ca2d485b94e6.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/6862-efff51af68016934.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/6874-c780af22003d1284.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/7393-567a776107bfae7f.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/822.f6b628d76038b708.js",revision:"f6b628d76038b708"},{url:"/_next/static/chunks/8321-a28475373aff6347.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/8330-5625b1ced2e903fb.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/8833-cfbcf13631972ae6.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/8844-408766209b8f1364.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/9397-0a7bf4d7c915ce40.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/9457-1956fbb2df6173cc.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/9558-a0d2ef8adeb8b898.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/ad2866b8.a559fc6296e7719c.js",revision:"a559fc6296e7719c"},{url:"/_next/static/chunks/app/(auth)/change-password/page-8c442023338eb4a6.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/(auth)/confirmation-token/page-1e6597f612fe1b13.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/(auth)/connexion/page-0e4e42ab61191774.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/(auth)/inscription/page-c3cdaf147e67e679.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/(auth)/layout-69cbccb3ecc6d736.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/(auth)/mot-de-passe-oublie/page-6bbd94b09f05cf90.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/(auth)/reset-password/page-cc74f821e1b699e3.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/_not-found/page-42d4bcc05786802f.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/a-propos-de-nous/page-a1a741e399a180ee.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/blogs/page-ce363ff47c659e42.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/category/page-b118363905e3d721.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/layout-a453b7a7f1c7ce08.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/orders/page-9a69d2787cb26766.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/page-6cda6e700ddc6b77.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/calendar/loading-6b8e1e9b1788032f.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/calendar/page-bf9b5b67e8c60e3c.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/layout-0d4cf458bb99197d.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/notifications/loading-15445a9dfe21540f.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/notifications/page-600f5915e4f59af8.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/page-8656454a87e87194.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/%5Bid%5D/page-8b1733fd11d1e4ed.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/avability/page-4bc6f3067134ec40.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/departments/page-1b41537871fba8c6.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/layout-6a61ee5e9b12c2fb.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/page-21f2b5cbaf6a4e7b.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/settings/page-c9cce6bf74c1e7f7.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/users/%5Bid%5D/page-58deff713438a9af.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/users/page-48ebd834f65c3a8a.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/waiting/page-42b4433a04ebcad6.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/blog/page-667de90d4dd11035.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/admin/page-9d5457edde39b459.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/auth/%5B...nextauth%5D/route-7a9ba1f3929f6180.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/change-password/route-f934bef0f2026d01.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/confirm-email/route-5fe2990d8dfd3d13.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/departments/route-ec98a60baa371c9c.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/generate-vapid-keys/route-78f060a3835f50a9.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/get-info-organisation/%5Bid%5D/route-c2de8abda26b1357.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/inngest/route-ad33c675cf9ddeff.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/appointments/%5BappointmentId%5D/route-23cd139a1e377c37.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/appointments/%5BappointmentId%5D/services/%5BserviceId%5D/complete/route-128f6d054e6ee45c.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/appointments/%5BappointmentId%5D/services/%5BserviceId%5D/start/route-a069012008a2759b.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/appointments/all/route-9ef0b2206988d651.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/appointments/barber/all/route-d6046c30d8786f54.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/appointments/barber/route-23e7263c7e61ae0d.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/appointments/date/route-7158166b8a5475c7.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/appointments/route-83deb18c59cf1804.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/availability/route-b969fcedd7bc95be.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/availability/slots/route-9009cfa35db5fe13.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/categories/%5Bid_category%5D/route-908a60b6df7d0626.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/categories/route-b41e10774ee158ac.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/data/route-4e24909ff90f2b4b.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/departments/%5BdepartmentId%5D/route-54c87230c0edc44a.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/departments/batch/route-9fc59b87d40070d9.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/departments/route-3040f41f1bf0593f.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/financial-chart/route-6eb636622ea00e72.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/membership/route-78201b05b71d12b2.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/notifications/%5Bid_notification%5D/route-a73c19988b1640a0.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/notifications/route-d52b1e52cbef1ee7.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/notifications/subscribe/route-3ac02f662473e998.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/orders/%5Bid_order%5D/route-8c09f81c90522961.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/orders/route-83ebbcc06982a390.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/products/%5Bid_product%5D/route-3232ef69076a6235.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/products/route-522f3253ef525783.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/route-d956aa7518001145.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/services/%5BserviceId%5D/assign/route-8eb60c56699c1c5f.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/services/%5BserviceId%5D/route-efb7f211402ed20d.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/services/route-30b6495f3a7806cd.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/settings/route-d6ab0ea74fc622f8.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/users/%5Buser_id%5D/route-692fcc0df0745d81.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/users/%5Buser_id%5D/services/route-f42710993c27acd5.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/users/%5Buser_id%5D/stats/route-e7927a470b9ab9b1.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/%5Bid%5D/users/route-84b710c3473307b1.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/departements/route-622b628a470ff8a8.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/organizations/route-26f9e7b7efa2606b.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/register/route-26bfd566906ebb8e.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/request-reset-password/route-c7587eb114378367.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/api/uploadthing/route-2596d4a5c5d723dd.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/business/layout-0802b300a2b05ac5.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/business/page-914f6d6d841e7589.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/layout-14151359143723dc.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/offline/page-9f14074ac31b1291.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/page-eec650677d2483c1.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/booking/confirmation/%5Bid_transaction%5D/page-79a29d453fbec11e.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/booking/page-3db55bc0c1276503.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/boutique/checkout/page-bf33537cd7b9f348.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/boutique/confirmation/%5BcommandeId%5D/page-368884b27c448a07.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/boutique/page-82b969f51c8197fe.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/page-80dd658d61c9e12c.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/queue/loading-03eb1be542cb94e8.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/queue/page-716dd5f0bf8564d8.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/bc98253f.fdffbd48dc845d22.js",revision:"fdffbd48dc845d22"},{url:"/_next/static/chunks/ca377847-b5180a0376487452.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/framework-82b67a6346ddd02b.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/main-1e16f04f764c9968.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/main-app-625daa935ca78554.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/pages/_app-5d1abe03d322390c.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/pages/_error-3b2a1d523de49635.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-5345bd88e0b171b2.js",revision:"nm3dlxk9zzwMUszZH5hi3"},{url:"/_next/static/css/afc566c2a9734ab2.css",revision:"afc566c2a9734ab2"},{url:"/_next/static/css/b3cbcd051438d1d5.css",revision:"b3cbcd051438d1d5"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/6d93bde91c0c2823-s.woff2",revision:"621a07228c8ccbfd647918f1021b4868"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/a34f9d1faa5f3315-s.p.woff2",revision:"d4fe31e6a2aebc06b8d6e558c9141119"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/nm3dlxk9zzwMUszZH5hi3/_buildManifest.js",revision:"8bbe8c260ec7a7058864bb8002c6ea4d"},{url:"/_next/static/nm3dlxk9zzwMUszZH5hi3/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/bg-1.png",revision:"bad3465ade687bbdb386ef821e5bcc27"},{url:"/bg-2.png",revision:"bdf522bc35108c7e32a74cfb2e0447ec"},{url:"/bg-3.png",revision:"d1a79bdcf7d00b936f0e255ee2495f7f"},{url:"/bg-4.png",revision:"193d7f5782969224807faec4461d767b"},{url:"/bg-5.png",revision:"c62ac8af8022ae40508359e0d5bcb486"},{url:"/bg.png",revision:"6923d2104ddacf4d4c27d3e1f80c593b"},{url:"/bijou.jpg",revision:"f3456dceeab33f7ce104e70fc25a546b"},{url:"/casquette.jpg",revision:"cc9fb2a69f69aeb12a00fc46b1d93fbf"},{url:"/coiffure.png",revision:"829e8db0636ec16153d38a0efd20bfb8"},{url:"/coloration.jpg",revision:"04f15f5072b55465eff50f2c2f22966a"},{url:"/epilation.jpg",revision:"1cf6f32712dbd0517e216229ac31264a"},{url:"/epilation.png",revision:"6b878a05082e1ff64d353f829148bfeb"},{url:"/epilation.webp",revision:"0024dedacfd2f31cc579700e8afffc53"},{url:"/extension.jpg",revision:"c19d99d57b1945e3d1ddc6cfd7c14f07"},{url:"/file.svg",revision:"d09f95206c3fa0bb9bd9fefabfd0ea71"},{url:"/globe.svg",revision:"2aaafa6a49b6563925fe440891e32717"},{url:"/huile.jpg",revision:"12b4e13ada73fc93a8d107841f866a13"},{url:"/icons/apple-icon-152x152.png",revision:"f886b3082f9458a21b729d3ed9a82c01"},{url:"/icons/apple-icon-167x167.png",revision:"5938c61d9f0be6e0379c7e355e7af803"},{url:"/icons/apple-icon-180x180.png",revision:"e8eb652f3422cd29a4e756e78c2b365a"},{url:"/icons/favicon-16x16.png",revision:"56662be1b6bb6ca3a81111cb2c78139c"},{url:"/icons/favicon-32x32.png",revision:"5db597461666cef61571bd993a580b5c"},{url:"/icons/icon-128x128.png",revision:"81cd2eea4fbcbd6d333156be193e8e12"},{url:"/icons/icon-144x144.png",revision:"db9806aa1958b9869c6a315c19da4db9"},{url:"/icons/icon-152x152.png",revision:"f886b3082f9458a21b729d3ed9a82c01"},{url:"/icons/icon-167x167.png",revision:"5938c61d9f0be6e0379c7e355e7af803"},{url:"/icons/icon-16x16.png",revision:"56662be1b6bb6ca3a81111cb2c78139c"},{url:"/icons/icon-180x180.png",revision:"e8eb652f3422cd29a4e756e78c2b365a"},{url:"/icons/icon-192x192.png",revision:"f8b78ace351778ac2576edfede232f90"},{url:"/icons/icon-32x32.png",revision:"5db597461666cef61571bd993a580b5c"},{url:"/icons/icon-384x384.png",revision:"88e02c5436cf8861d1b6a76c0c917788"},{url:"/icons/icon-512x512.png",revision:"33510ffd245548b59088d4ea417474c1"},{url:"/icons/icon-72x72.png",revision:"40fc43ebc6f2146cb7b75d20265ea017"},{url:"/icons/icon-96x96.png",revision:"be6b15d47c9384b239c0510049283e6f"},{url:"/logo-black.png",revision:"5621943d320617ace5d02251bc8f48c8"},{url:"/logo-white.png",revision:"3aa9f9af518149021a7fa4c52836b99d"},{url:"/logo.png",revision:"6ff664a8bdebc8253129b6911220e53d"},{url:"/mackup.svg",revision:"a45fc1353a313a5628ce40e2cf5b4d1e"},{url:"/maintien.jpg",revision:"976de7c89e967f1109faae3da762009d"},{url:"/manifest.json",revision:"6c958fda6323b29ff18d0fccf725050f"},{url:"/marley.jpg",revision:"d5936ebd5e268d6258cc0cc4c0ee68df"},{url:"/massage.png",revision:"d3970a316a3a95164b06d3317f907062"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/offline-image.png",revision:"511ae0f5c05d6dfa788756a677f68ce6"},{url:"/pedicure.png",revision:"47600f7eb034f5477f81a562b3c11187"},{url:"/placeholder.svg",revision:"35707bd9960ba5281c72af927b79291f"},{url:"/reparation.jpg",revision:"74be29d4536becd1f397bf22a80192eb"},{url:"/reprises.jpg",revision:"a848f04251b24708d3b61451733aeb6b"},{url:"/saloon.png",revision:"d25d79404164f106d59011663aaec48b"},{url:"/saloona 2.png",revision:"3aa9f9af518149021a7fa4c52836b99d"},{url:"/saloona.png",revision:"192e9e654fc0548ef294700b8e57b69c"},{url:"/service-worker.js",revision:"cbb9dc8970371d66f659c3bfde8d021b"},{url:"/shampion.jpeg",revision:"7ffe5fde704e4160de868a7f063d5537"},{url:"/start.jpeg",revision:"dffd12b9900eddb0779915c657d685a0"},{url:"/tee-shirt.jpg",revision:"ef044e0cde3caf1a7d565751bf966440"},{url:"/twist.jpg",revision:"44d92021d7be346abd81ab01700710b8"},{url:"/vercel.svg",revision:"c0af2f507b369b085b35ef4bbe3bcf1e"},{url:"/visage.png",revision:"a47f0e117e5912c2ff759d59881e4ca6"},{url:"/window.svg",revision:"a2760511c65806022ad20adf74370ff3"}],{ignoreURLParametersMatching:[]}),i.cleanupOutdatedCaches(),i.registerRoute("/",new i.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:i,response:s,event:e,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),i.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new i.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new i.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),i.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new i.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new i.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),i.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new i.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new i.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),i.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new i.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new i.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),i.registerRoute(/\/_next\/image\?url=.+$/i,new i.StaleWhileRevalidate({cacheName:"next-image",plugins:[new i.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),i.registerRoute(/\.(?:mp3|wav|ogg)$/i,new i.CacheFirst({cacheName:"static-audio-assets",plugins:[new i.RangeRequestsPlugin,new i.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),i.registerRoute(/\.(?:mp4)$/i,new i.CacheFirst({cacheName:"static-video-assets",plugins:[new i.RangeRequestsPlugin,new i.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),i.registerRoute(/\.(?:js)$/i,new i.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new i.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),i.registerRoute(/\.(?:css|less)$/i,new i.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new i.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),i.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new i.StaleWhileRevalidate({cacheName:"next-data",plugins:[new i.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),i.registerRoute(/\.(?:json|xml|csv)$/i,new i.NetworkFirst({cacheName:"static-data-assets",plugins:[new i.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),i.registerRoute((({url:i})=>{if(!(self.origin===i.origin))return!1;const s=i.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new i.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new i.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),i.registerRoute((({url:i})=>{if(!(self.origin===i.origin))return!1;return!i.pathname.startsWith("/api/")}),new i.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new i.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),i.registerRoute((({url:i})=>!(self.origin===i.origin)),new i.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new i.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
+if(!self.define){let a,e={};const s=(s,i)=>(s=new URL(s+".js",i).href,e[s]||new Promise((e=>{if("document"in self){const a=document.createElement("script");a.src=s,a.onload=e,document.head.appendChild(a)}else a=s,importScripts(s),e()})).then((()=>{let a=e[s];if(!a)throw new Error(`Module ${s} didn’t register its module`);return a})));self.define=(i,t)=>{const n=a||("document"in self?document.currentScript.src:"")||location.href;if(e[n])return;let c={};const o=a=>s(a,n),d={module:{uri:n},exports:c,require:o};e[n]=Promise.all(i.map((a=>d[a]||o(a)))).then((a=>(t(...a),c)))}}define(["./workbox-4754cb34"],(function(a){"use strict";importScripts(),self.skipWaiting(),a.clientsClaim(),a.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"cfcbed38dcb594759428fe6ab9377edb"},{url:"/_next/static/chunks/1119-ed7d0dc84b127bcb.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/124.992a6e622f731d4f.js",revision:"992a6e622f731d4f"},{url:"/_next/static/chunks/1375-26d9148a57deac3e.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/1581.9ec8cf16f255b7d2.js",revision:"9ec8cf16f255b7d2"},{url:"/_next/static/chunks/164f4fb6-77124367fedc6fd1.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/1880-b152ec88ad675b5b.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/202-69de734f4a0b6b19.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/231-eed28fb385e310b9.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/2514-4b4938523cf82999.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/2747-ff64637a75ddfe46.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/2808-3adf8b8f8bd517c7.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/323-cc8ed5151cd235c0.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/329-61c9ee602f8b0ad1.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/3356-fb1e9336135a54ec.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/3454-6696efa4f51c3330.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/3701-55075a5182387eb9.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/381-203f9e54a23f6be6.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/4298-6419c90fa1eec006.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/462-00ddbc01a6437ef5.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/4723-4d43a4a0c0f744b4.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/508-136943f7964faf35.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/5129-700259b9467c7df5.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/5737-beaa9f25856c3efa.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/600-58691f2d7c35c42e.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/602dbae6-1523a36c3ae7ba0c.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/6343-ed8bda79dcad562e.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/6388-43cd87e389cca48b.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/6619-247209edd18b0ed3.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/6827-f49d28a07ce8b20b.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/6963-bcd8d2b7a5e42ddd.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/7023-4d9c74e33c73c619.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/7032-97d714e3a538c087.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/7207-0f45cc51027a52b4.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/7717-3656d41720bb2dc5.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/7776-26ea0ab1c8c5043d.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/8173-b25e00649cbd6f0c.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/8277-f10d058bb39127ad.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/8322-e5aab8ab1727dfa0.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/8972-79195a8bc1739f6b.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/9134-696624a7614439e0.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/9597-22c347f1298c9df4.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/9637-8046e8ca918e4d97.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/9737-9a9289192aa3c252.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/9783-ee55fc608dcbebad.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/ad2866b8.0a10d661ca31722a.js",revision:"0a10d661ca31722a"},{url:"/_next/static/chunks/app/(auth)/change-password/page-b7d35acf7e3272de.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/(auth)/confirmation-token/page-f359ad8d3cd8995e.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/(auth)/connexion/page-49dd49a5771f5209.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/(auth)/inscription/page-6267005a35e81769.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/(auth)/layout-7473a864188993cc.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/(auth)/mot-de-passe-oublie/page-b2b7da775af7f0f3.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/(auth)/reset-password/page-6ce77238e4cc4da1.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/_not-found/page-ed75635ea4b25f64.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/a-propos-de-nous/page-4a0c19eef212ae17.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/blogs/page-7f47e485c9315862.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/category/page-8aacd516dca2b220.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/layout-35d9dc950b27d878.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/orders/page-a144c0b4690177ab.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/boutique/page-8af62c83f352e406.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/calendar/loading-596f004cb4d2bab3.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/calendar/page-9d46dfe1b30303f6.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/layout-6b277599a4abaf2a.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/notifications/loading-fee67f8cea0882b7.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/notifications/page-79dbbdf8f88c0bee.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/page-1ab68de932b66f55.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/%5Bid%5D/page-7107fc9f4825ad48.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/avability/page-8bea5fda39d33a08.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/departments/page-a667ffbf1cd5f14a.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/layout-3e2c19b91f497038.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/services/page-03294b482932f1c9.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/settings/page-19813672c8f1996e.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/users/%5Bid%5D/page-4e2b7515d6cac829.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/users/page-8ff16ded262ede68.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/%5Bsalon_id%5D/waiting/page-616cd0315022e811.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/blog/page-05cd4c78e48640f1.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/admin/page-57b6990587c3a402.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/business/layout-804a5323e6057e17.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/business/page-59aaebf08641fd13.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/layout-500a8272a57d5beb.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/offline/page-5e9c8b3564b9aea8.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/page-4f76210df67088fd.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/booking/confirmation/%5Bid_transaction%5D/page-8dcac3818f1a6cd6.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/booking/page-fd7acf579ad5c8ce.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/boutique/checkout/page-a15b6179cfc3b94c.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/boutique/confirmation/%5BcommandeId%5D/page-95f99517473e40fa.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/boutique/page-88e1083208887727.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/page-3e885ab45ff941d4.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/queue/loading-08bebf99ea9919aa.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/app/salon/%5Bid%5D/queue/page-abce564321515aa0.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/bc98253f.da2689c4a01c0e56.js",revision:"da2689c4a01c0e56"},{url:"/_next/static/chunks/ca377847-a55438354b0797ad.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/fd9d1056-b2ff4dea245f9be6.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/framework-08aa667e5202eed8.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/main-app-87d383141b4b3d52.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/main-f09965453c48c295.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/pages/_app-f870474a17b7f2fd.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/pages/_error-c66a4e8afc46f17b.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-e2b68995adf1a577.js",revision:"d9BtGajIBCqvXoysztaPz"},{url:"/_next/static/css/48340cfe5c3cfec7.css",revision:"48340cfe5c3cfec7"},{url:"/_next/static/css/8756cc1a48e5f022.css",revision:"8756cc1a48e5f022"},{url:"/_next/static/d9BtGajIBCqvXoysztaPz/_buildManifest.js",revision:"3e2d62a10f4d6bf0b92e14aecf7836f4"},{url:"/_next/static/d9BtGajIBCqvXoysztaPz/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/6d93bde91c0c2823-s.woff2",revision:"621a07228c8ccbfd647918f1021b4868"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/a34f9d1faa5f3315-s.p.woff2",revision:"d4fe31e6a2aebc06b8d6e558c9141119"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/bg-1.png",revision:"bad3465ade687bbdb386ef821e5bcc27"},{url:"/bg-2.png",revision:"bdf522bc35108c7e32a74cfb2e0447ec"},{url:"/bg-3.png",revision:"d1a79bdcf7d00b936f0e255ee2495f7f"},{url:"/bg-4.png",revision:"193d7f5782969224807faec4461d767b"},{url:"/bg-5.png",revision:"c62ac8af8022ae40508359e0d5bcb486"},{url:"/bg.png",revision:"6923d2104ddacf4d4c27d3e1f80c593b"},{url:"/bijou.jpg",revision:"f3456dceeab33f7ce104e70fc25a546b"},{url:"/casquette.jpg",revision:"cc9fb2a69f69aeb12a00fc46b1d93fbf"},{url:"/coiffure.png",revision:"829e8db0636ec16153d38a0efd20bfb8"},{url:"/coloration.jpg",revision:"04f15f5072b55465eff50f2c2f22966a"},{url:"/epilation.jpg",revision:"1cf6f32712dbd0517e216229ac31264a"},{url:"/epilation.png",revision:"6b878a05082e1ff64d353f829148bfeb"},{url:"/epilation.webp",revision:"0024dedacfd2f31cc579700e8afffc53"},{url:"/extension.jpg",revision:"c19d99d57b1945e3d1ddc6cfd7c14f07"},{url:"/file.svg",revision:"d09f95206c3fa0bb9bd9fefabfd0ea71"},{url:"/globe.svg",revision:"2aaafa6a49b6563925fe440891e32717"},{url:"/huile.jpg",revision:"12b4e13ada73fc93a8d107841f866a13"},{url:"/icons/apple-icon-152x152.png",revision:"f886b3082f9458a21b729d3ed9a82c01"},{url:"/icons/apple-icon-167x167.png",revision:"5938c61d9f0be6e0379c7e355e7af803"},{url:"/icons/apple-icon-180x180.png",revision:"e8eb652f3422cd29a4e756e78c2b365a"},{url:"/icons/favicon-16x16.png",revision:"56662be1b6bb6ca3a81111cb2c78139c"},{url:"/icons/favicon-32x32.png",revision:"5db597461666cef61571bd993a580b5c"},{url:"/icons/icon-128x128.png",revision:"81cd2eea4fbcbd6d333156be193e8e12"},{url:"/icons/icon-144x144.png",revision:"db9806aa1958b9869c6a315c19da4db9"},{url:"/icons/icon-152x152.png",revision:"f886b3082f9458a21b729d3ed9a82c01"},{url:"/icons/icon-167x167.png",revision:"5938c61d9f0be6e0379c7e355e7af803"},{url:"/icons/icon-16x16.png",revision:"56662be1b6bb6ca3a81111cb2c78139c"},{url:"/icons/icon-180x180.png",revision:"e8eb652f3422cd29a4e756e78c2b365a"},{url:"/icons/icon-192x192.png",revision:"f8b78ace351778ac2576edfede232f90"},{url:"/icons/icon-32x32.png",revision:"5db597461666cef61571bd993a580b5c"},{url:"/icons/icon-384x384.png",revision:"88e02c5436cf8861d1b6a76c0c917788"},{url:"/icons/icon-512x512.png",revision:"33510ffd245548b59088d4ea417474c1"},{url:"/icons/icon-72x72.png",revision:"40fc43ebc6f2146cb7b75d20265ea017"},{url:"/icons/icon-96x96.png",revision:"be6b15d47c9384b239c0510049283e6f"},{url:"/logo-black.png",revision:"5621943d320617ace5d02251bc8f48c8"},{url:"/logo-white.png",revision:"3aa9f9af518149021a7fa4c52836b99d"},{url:"/logo.png",revision:"6ff664a8bdebc8253129b6911220e53d"},{url:"/mackup.svg",revision:"a45fc1353a313a5628ce40e2cf5b4d1e"},{url:"/maintien.jpg",revision:"976de7c89e967f1109faae3da762009d"},{url:"/manifest.json",revision:"6c958fda6323b29ff18d0fccf725050f"},{url:"/marley.jpg",revision:"d5936ebd5e268d6258cc0cc4c0ee68df"},{url:"/massage.png",revision:"d3970a316a3a95164b06d3317f907062"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/offline-image.png",revision:"511ae0f5c05d6dfa788756a677f68ce6"},{url:"/pedicure.png",revision:"47600f7eb034f5477f81a562b3c11187"},{url:"/placeholder.svg",revision:"35707bd9960ba5281c72af927b79291f"},{url:"/reparation.jpg",revision:"74be29d4536becd1f397bf22a80192eb"},{url:"/reprises.jpg",revision:"a848f04251b24708d3b61451733aeb6b"},{url:"/saloon.png",revision:"d25d79404164f106d59011663aaec48b"},{url:"/saloona 2.png",revision:"3aa9f9af518149021a7fa4c52836b99d"},{url:"/saloona.png",revision:"192e9e654fc0548ef294700b8e57b69c"},{url:"/service-worker.js",revision:"cbb9dc8970371d66f659c3bfde8d021b"},{url:"/shampion.jpeg",revision:"7ffe5fde704e4160de868a7f063d5537"},{url:"/start.jpeg",revision:"dffd12b9900eddb0779915c657d685a0"},{url:"/tee-shirt.jpg",revision:"ef044e0cde3caf1a7d565751bf966440"},{url:"/twist.jpg",revision:"44d92021d7be346abd81ab01700710b8"},{url:"/vercel.svg",revision:"c0af2f507b369b085b35ef4bbe3bcf1e"},{url:"/visage.png",revision:"a47f0e117e5912c2ff759d59881e4ca6"},{url:"/window.svg",revision:"a2760511c65806022ad20adf74370ff3"}],{ignoreURLParametersMatching:[]}),a.cleanupOutdatedCaches(),a.registerRoute("/",new a.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:a,response:e,event:s,state:i})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),a.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new a.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new a.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),a.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new a.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new a.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),a.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new a.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new a.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),a.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new a.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new a.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),a.registerRoute(/\/_next\/image\?url=.+$/i,new a.StaleWhileRevalidate({cacheName:"next-image",plugins:[new a.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),a.registerRoute(/\.(?:mp3|wav|ogg)$/i,new a.CacheFirst({cacheName:"static-audio-assets",plugins:[new a.RangeRequestsPlugin,new a.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),a.registerRoute(/\.(?:mp4)$/i,new a.CacheFirst({cacheName:"static-video-assets",plugins:[new a.RangeRequestsPlugin,new a.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),a.registerRoute(/\.(?:js)$/i,new a.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new a.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),a.registerRoute(/\.(?:css|less)$/i,new a.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new a.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),a.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new a.StaleWhileRevalidate({cacheName:"next-data",plugins:[new a.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),a.registerRoute(/\.(?:json|xml|csv)$/i,new a.NetworkFirst({cacheName:"static-data-assets",plugins:[new a.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),a.registerRoute((({url:a})=>{if(!(self.origin===a.origin))return!1;const e=a.pathname;return!e.startsWith("/api/auth/")&&!!e.startsWith("/api/")}),new a.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new a.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),a.registerRoute((({url:a})=>{if(!(self.origin===a.origin))return!1;return!a.pathname.startsWith("/api/")}),new a.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new a.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),a.registerRoute((({url:a})=>!(self.origin===a.origin)),new a.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new a.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
